content/en/blog/2023/logs-collection/yoda/requirements.txt
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
